+++ CHANGELOG.md
@@ -607,3 +607,4 @@
 - [2026-04-23 18:38] fix(auth): prevent double tap duplicate action in auth
 - [2026-04-23 18:53] feat(storage): add async storage stream provider
 - [2026-04-23 18:55] style(forms): adjust design tokens, typography, and contrast in forms
+- [2026-04-23 18:55] test(test): add unit tests for test business logic